REFACTOR : Delete  AwsConfg, S3Service File,    Add AwsConfg.java, S3Service.Java
