+++ src/main/java/sopt/univoice/domain/auth/dto/CheckEmailRequest.java
@@ -3,6 +3,7 @@ package sopt.univoice.domain.auth.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+
 @Getter
 @NoArgsConstructor
 public class CheckEmailRequest {
