+++ src/main/java/sopt/univoice/domain/auth/service/AuthService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import sopt.univoice.domain.auth.dto.CheckEmailRequest;
 import sopt.univoice.domain.auth.repository.AuthRepository;
+import sopt.univoice.infra.external.S3Service;
 
 @Service
 @RequiredArgsConstructor
@@ -11,6 +12,7 @@ public class AuthService {
 
 
     private final AuthRepository authRepository;
+    private final S3Service s3Service;
 
     public boolean isDuplicateEmail(CheckEmailRequest checkEmailRequest) {
         return authRepository.existsByEmail(checkEmailRequest.getEmail());
