+++ src/main/java/sopt/univoice/infra/config/AwsConfig.java
@@ -1,4 +1,4 @@
-package sopt.univoice.external;
+package sopt.univoice.infra.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
