+++ src/main/java/sopt/univoice/infra/external/S3Service.java
@@ -1,4 +1,5 @@
-package sopt.univoice.external;
+package sopt.univoice.infra.external;
+
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
@@ -7,6 +8,7 @@ import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
+import sopt.univoice.infra.config.AwsConfig;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -78,3 +80,4 @@ public class S3Service {
     }
 
 }
+
